backend/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
google-cloud-firestore==2.0.2
google-cloud-ndb==1.7.2
google-cloud-storage==1.35.0
